Refactor Expanse model and related components: Change amount type from int to double for better precision, update serialization methods, and adjust UI to reflect changes in expense amount handling.

+++ lib/screens/add_expense/add_expanse.screen.dart
@@ -20,7 +20,7 @@ class AddExpanseScreens extends StatelessWidget {
             // You can integrate with your BLoC here:
             context.read<ExpanseBloc>().add(
               CreateExpenseEvent(
-                Expanse(name: title, amount: amount.toInt(), dayName: day),
+                Expanse(name: title, amount: amount, dayName: day),
               ),
             );
